Restructure to control method-parsing complexity
* How did Jacob fit it into 4 files? He must be the Dr. Cerny of
  code-categorization, but with files instead of lectures.
