+++ src/lang/java/class_def.rs
@@ -1,6 +1,6 @@
 use crate::java::method_def::parse_method;
 use crate::java::modifier::{find_modifier, parse_modifiers, Modifier};
-use crate::java::vartype::find_type;
+use crate::java::util::vartype::find_type;
 use crate::parse::AST;
 use crate::prophet::*;
