+++ src/lang/java/util/mod.rs
@@ -1,5 +1,7 @@
 use crate::AST;
 
+pub(crate) mod vartype;
+
 /// Convert a vector into an Option. If the vector is empty, swaps it out for None; otherwise is Some(vector)
 pub(crate) fn fold_vec<T>(vector: Vec<T>) -> Option<Vec<T>> {
     if !vector.is_empty() {
@@ -16,9 +18,4 @@ pub(crate) fn stringify_tree_children(ast: &AST) -> String {
         buffer.push_str(&*member.value);
     }
     buffer
-}
-
-/// Logs an unknown tag was encountered. You better not think too much about that.
-pub(crate) fn log_unknown_tag(tag: &str, parent: &str) {
-    eprintln!("Unknown tag {} encountered while parsing {}!", tag, parent);
 }
